Remove decoder_cell from SeqDecoder, rename SimpleSeqDecoder -> RnnSeqDecoder
