+++ allennlp/modules/seq2seq_decoders/__init__.py
@@ -8,3 +8,4 @@ The available Seq2Seq decoders are
 
 """
 from allennlp.modules.seq2seq_decoders.lstm_decoder_cell import LstmDecoderCell
+from allennlp.modules.seq2seq_decoders.rnn_seq_decoder import RnnSeqDecoder
