+++ allennlp/modules/seq2seq_decoders/rnn_seq_decoder.py
@@ -1,18 +1,55 @@
 from typing import Dict, List, Tuple
 
-import numpy
 import torch
 import torch.nn.functional as F
+from torch.nn import Linear
 
-from allennlp.data import Vocabulary
-from allennlp.modules.seq2seq_decoders.decoder_cell import DecoderCell
+from allennlp.common import Registrable
 from allennlp.modules.seq2seq_decoders.seq_decoder import SeqDecoder
+from allennlp.data import Vocabulary
+from allennlp.modules import Embedding
+from allennlp.modules.seq2seq_decoders.decoder_cell import DecoderCell
 from allennlp.nn import util
+from allennlp.nn.beam_search import BeamSearch
 from allennlp.training.metrics import Metric
 
 
-@SeqDecoder.register("simple_seq_decoder")
-class SimpleSeqDecoder(SeqDecoder):
+@SeqDecoder.register("rnn_seq_decoder")
+class RnnSeqDecoder(SeqDecoder):
+    """
+
+    Parameters
+    ----------
+    vocab : ``Vocabulary``, required
+        Vocabulary containing source and target vocabularies. They may be under the same namespace
+        (`tokens`) or the target tokens can have a different namespace, in which case it needs to
+        be specified as `target_namespace`.
+    decoder_cell : ``DecoderCell``, required
+        Module that contains implementation of neural network for decoding output elements
+    max_decoding_steps : ``int``
+        Maximum length of decoded sequences.
+    bidirectional_input : ``bool``
+        If input encoded sequence was produced by bidirectional encoder.
+        If True, the first encode step of back direction will be used as initial hidden state.
+        If not, the will be used the last step only.
+    target_namespace : ``str``, optional (default = 'target_tokens')
+        If the target side vocabulary is different from the source side's, you need to specify the
+        target's namespace here. If not, we'll assume it is "tokens", which is also the default
+        choice for the source side, and this might cause them to share vocabularies.
+    beam_size : ``int``, optional (default = None)
+        Width of the beam for beam search. If not specified, greedy decoding is used.
+    tensor_based_metric : ``Metric``, optional (default = BLEU)
+        A metric to track on validation data that takes raw tensors when its called.
+        This metric must accept two arguments when called: a batched tensor
+        of predicted token indices, and a batched tensor of gold token indices.
+    token_based_metric : ``Metric``, optional (default = None)
+        A metric to track on validation data that takes lists of lists of tokens
+        as input. This metric must accept two arguments when called, both
+        of type `List[List[str]]`. The first is a predicted sequence for each item
+        in the batch and the second is a gold sequence for each item in the batch.
+    scheduled_sampling_ratio : ``float``
+        Defines ratio between teacher forced training and real output usage.
+    """
 
     def __init__(
             self,
@@ -21,23 +58,54 @@ class SimpleSeqDecoder(SeqDecoder):
             max_decoding_steps: int,
             bidirectional_input: bool = False,
             beam_size: int = None,
-            target_embedding_dim: int = None,
             target_namespace: str = "tokens",
             tensor_based_metric: Metric = None,
             token_based_metric: Metric = None,
             scheduled_sampling_ratio: float = 0.,
     ):
-        super().__init__(
+
+        super(RnnSeqDecoder, self).__init__(
             vocab=vocab,
-            decoder_cell=decoder_cell,
-            max_decoding_steps=max_decoding_steps,
-            bidirectional_input=bidirectional_input,
-            beam_size=beam_size,
-            target_embedding_dim=target_embedding_dim,
             target_namespace=target_namespace,
             tensor_based_metric=tensor_based_metric,
-            token_based_metric=token_based_metric,
+            token_based_metric=token_based_metric
         )
+
+        self.bidirectional_input = bidirectional_input
+
+        self._target_vocab_size = self.vocab.get_vocab_size(self._target_namespace)
+
+        # At prediction time, we use a beam search to find the most likely sequence of target tokens.
+        beam_size = beam_size or 1
+        self._max_decoding_steps = max_decoding_steps
+        self._beam_search = BeamSearch(self._end_index, max_steps=max_decoding_steps, beam_size=beam_size)
+
+        # Decodes the sequence of encoded hidden states into e new sequence of hidden states.
+        self.decoder_cell = decoder_cell
+
+        # Decoder output dim needs to be the same as the encoder output dim since we initialize the
+        # hidden state of the decoder with the final hidden state of the encoder.
+        # We arbitrarily set the decoder's input dimension to be the same as the output dimension.
+        self.decoder_output_dim = self.decoder_cell.get_output_dim()
+        self.decoder_input_dim = self.decoder_output_dim
+        self.encoder_output_dim = self.decoder_input_dim
+
+        target_vocab_size = self.vocab.get_vocab_size(self._target_namespace)
+
+        # The decoder input will be a function of the embedding of the previous predicted token,
+        # an attended encoder hidden state called the "attentive read", and another
+        # weighted sum of the encoder hidden state called the "selective read".
+        # While the weights for the attentive read are calculated by an `Attention` module,
+        # the weights for the selective read are simply the predicted probabilities
+        # corresponding to each token in the source sentence that matches the target
+        # token from the previous timestep.
+
+        # Dense embedding of vocab words in the target space.
+        self._target_embedder = Embedding(target_vocab_size, self.decoder_cell.target_embedding_dim)
+
+        # We project the hidden state from the decoder into the output vocabulary space
+        # in order to get log probabilities of each target token, at each time step.
+        self._output_projection_layer = Linear(self.decoder_cell.get_output_dim(), target_vocab_size)
 
         self._scheduled_sampling_ratio = scheduled_sampling_ratio
 
@@ -251,35 +319,26 @@ class SimpleSeqDecoder(SeqDecoder):
 
         return util.sequence_cross_entropy_with_logits(logits, relevant_targets, relevant_mask)
 
-    def decode(self, output_dict: Dict[str, torch.Tensor]) -> Dict[str, torch.Tensor]:
+    def _init_decoder_state(self, state: Dict[str, torch.Tensor]) -> Dict[str, torch.Tensor]:
         """
-        Finalize predictions.
-
-        This method overrides ``Model.decode``, which gets called after ``Model.forward``, at test
-        time, to finalize predictions. The logic for the decoder part of the encoder-decoder lives
-        within the ``forward`` method.
-
-        This method trims the output predictions to the first end symbol, replaces indices with
-        corresponding tokens, and adds a field called ``predicted_tokens`` to the ``output_dict``.
+        Initialize the encoded state to be passed to the first decoding time step.
         """
-        predicted_indices = output_dict["predictions"]
-        if not isinstance(predicted_indices, numpy.ndarray):
-            predicted_indices = predicted_indices.detach().cpu().numpy()
-        all_predicted_tokens = []
-        for indices in predicted_indices:
-            # Beam search gives us the top k results for each source sentence in the batch
-            # but we just want the single best.
-            if len(indices.shape) > 1:
-                indices = indices[0]
-            indices = list(indices)
-            # Collect indices till the first end_symbol
-            if self._end_index in indices:
-                indices = indices[:indices.index(self._end_index)]
-            predicted_tokens = [self.vocab.get_token_from_index(x, namespace=self._target_namespace)
-                                for x in indices]
-            all_predicted_tokens.append(predicted_tokens)
-        output_dict["predicted_tokens"] = all_predicted_tokens
-        return output_dict
+        batch_size, _ = state["source_mask"].size()
+
+        # Initialize the decoder hidden state with the final output of the encoder,
+        # and the decoder context with zeros.
+        # shape: (batch_size, encoder_output_dim)
+        final_encoder_output = util.get_final_encoder_states(
+            state["encoder_outputs"],
+            state["source_mask"],
+            bidirectional=self.bidirectional_input)
+
+        state.update(self.decoder_cell.init_decoder_state(batch_size, final_encoder_output))
+
+        return state
+
+    def get_output_dim(self):
+        return self.decoder_cell.get_output_dim()
 
     def take_step(self,
                   last_predictions: torch.Tensor,
@@ -321,3 +380,41 @@ class SimpleSeqDecoder(SeqDecoder):
         class_log_probabilities = F.log_softmax(output_projections, dim=-1)
 
         return class_log_probabilities, state
+
+    def forward(self,
+                encoder_out: Dict[str, torch.LongTensor],
+                target_tokens: Dict[str, torch.LongTensor] = None):
+
+        state = encoder_out
+
+        if target_tokens:
+            state = self._init_decoder_state(encoder_out)
+            # The `_forward_loop` decodes the input sequence and computes the loss during training
+            # and validation.
+            output_dict = self._forward_loss(state, target_tokens)
+        else:
+            output_dict = {}
+
+        if not self.training:
+            state = self._init_decoder_state(state)
+            predictions = self._forward_beam_search(state)
+            output_dict.update(predictions)
+
+            if target_tokens:
+                if self._tensor_based_metric is not None:
+                    # shape: (batch_size, beam_size, max_sequence_length)
+                    top_k_predictions = output_dict["predictions"]
+                    # shape: (batch_size, max_predicted_sequence_length)
+                    best_predictions = top_k_predictions[:, 0, :]
+                    # shape: (batch_size, target_sequence_length)
+
+                    self._tensor_based_metric(best_predictions, target_tokens["tokens"])  # type: ignore
+
+                if self._token_based_metric is not None:
+                    output_dict = self.decode(output_dict)
+                    predicted_tokens = output_dict['predicted_tokens']
+
+                    self._token_based_metric(predicted_tokens,  # type: ignore
+                                             [y.text for y in target_tokens["tokens"][1:-1]])
+
+        return output_dict
